requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
tenacity==8.2.3
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.3
loguru==0.7.2
